Adding Blackwood - Silent Halls and MapID support
Silent Halls has multiple maps that all use the same map name using the current code implementation.  That causes an issue where pins intended for different maps in this zone appear on all maps in this zone.  This adds support for Map IDs to fix the issue.  May consider using Map IDs as the default identifier for future updates, not sure yet.